Linux/LabJackScaffold fixes for logging bytes, casting to ints.
Bytes stream as chars, but the data being held in these bytes are
actually small unsigned integers.

+++ SurgSim/Devices/LabJack/linux/LabJackScaffold.cpp
@@ -170,8 +170,8 @@ bool readAndCheck(const LJ_HANDLE rawHandle, std::array<BYTE, LabJack::MAXIMUM_B
 	else if ((*readBytes)[6] != 0)
 	{
 		SURGSIM_LOG_SEVERE(logger) << "Failed to read response of " << text << " a device named '" << name <<
-			"'.  The device library returned an error code: " << (*readBytes)[6] << ", for frame: " <<
-			(*readBytes)[7] << std::endl << "  labjackusb error code: " << errno << "." << std::endl;
+			"'.  The device library returned an error code: " << static_cast<int>((*readBytes)[6]) << ", for frame: " <<
+			static_cast<int>((*readBytes)[7]) << std::endl << "  labjackusb error code: " << errno << "." << std::endl;
 		result = false;
 	}
 
@@ -914,8 +914,9 @@ bool LabJackScaffold::updateDevice(LabJackScaffold::DeviceData* info)
 		{
 			SURGSIM_LOG_SEVERE(m_logger) <<
 				"Failed to read response of " << errorText << " a device named '" <<	device->getName() <<
-				"'.  The number of words in the response is wrong.  Expected: " << dataWords << ".  Received: " <<
-				readBytes[2] << "." << std::endl << "  labjackusb error code: " << errno << "." << std::endl;
+				"'.  The number of words in the response is wrong.  Expected: " << static_cast<int>(dataWords) <<
+				".  Received: " << static_cast<int>(readBytes[2]) << "." << std::endl << "  labjackusb error code: " <<
+				errno << "." << std::endl;
 			result = false;
 		}
 	}
@@ -1306,7 +1307,7 @@ bool LabJackScaffold::configureTimers(DeviceData* deviceData)
 		{
 			SURGSIM_LOG_SEVERE(m_logger) << "Failed to read response of " << errorText << " a device named '" <<
 				device->getName() << "'.  The number of words in the response is wrong.  Expected: " <<
-				dataWords << ".  Received: " << readBytes[2] << "." << std::endl <<
+				static_cast<int>(dataWords) << ".  Received: " << static_cast<int>(readBytes[2]) << "." << std::endl <<
 				"  labjackusb error code: " << errno << "." << std::endl;
 			result = false;
 		}
@@ -1395,7 +1396,8 @@ bool LabJackScaffold::configureAnalog(DeviceData* deviceData)
 				SURGSIM_LOG_SEVERE(m_logger) << "Failed to read response of " << errorText << " a device named '" <<
 					device->getName() << "'.  The command bytes are wrong.  Expected byte 2: " <<
 					static_cast<int>(parameters.calibrationThirdByte[device->getModel()]) << ".  Received: " <<
-					readBytes[2] << "." << std::endl << "  labjackusb error code: " << errno << "." << std::endl;
+					static_cast<int>(readBytes[2]) << "." << std::endl << "  labjackusb error code: " << errno << "." <<
+					std::endl;
 				result = false;
 			}
 
@@ -1491,8 +1493,8 @@ bool LabJackScaffold::configureDigital(DeviceData* deviceData)
 			{
 				SURGSIM_LOG_SEVERE(m_logger) << "Failed to read response of " << errorText << " a device named '" <<
 					device->getName() << "'.  The number of words in the response is wrong.  Expected: " <<
-					dataWords << ".  Received: " << readBytes[2] << "." << std::endl <<
-					"  labjackusb error code: " << errno << "." << std::endl;
+					static_cast<int>(dataWords) << ".  Received: " << static_cast<int>(readBytes[2]) << "." <<
+					std::endl << "  labjackusb error code: " << errno << "." << std::endl;
 				result = false;
 			}
 		}
